Refactor notification handling in MainActivity and introduce NotificationIntentHandler
- Consolidated notification intent processing into a new NotificationIntentHandler class for better organization and maintainability.
- Updated MainActivity to utilize the new handler for both initial and new intents, streamlining the notification handling logic.
- Removed redundant code related to video conference and regular notification handling from MainActivity.
